requirements.txt
streamlit
requests
pydantic
python-dotenv
pypdf
groq
transformers
langchain
langchain-community
qdrant-client
langchain-groq
tensorflow
tf-keras
